+++ src/test/resources/edgeCases/updateWithQuotesInConcatenatedValue.php
@@ -1,0 +1,9 @@
+<?php
+
+class SomeClass {
+    public static function someMethod() { return ''; }
+}
+
+DB::table('users')
+    ->where('email', '!=', '"' . SomeClass::someMethod() . '"')
+    ->update(['email' => '"' .SomeClass::someMethod()  . '"']);