+++ src/test/resources/edgeCases/updateWithQuotesInValue.php
@@ -1,0 +1,5 @@
+<?php
+
+DB::table('users')
+    ->where('email', '!=', '"test"')
+    ->update(['email' => '"test"']);